Fix GCC warning
Array designated initializers may be valid in C99, but remain invalid in C++,
even C++20. GCC accepts them as an extension, hence the pedantic warning:
	warning: ISO C++ does not allow C99 designated initializers [-Wpedantic]

Because this is not the typical use case for array designated initializers,
remove them for clarity. Instead explain what the indices must be kept in
sync with the relevant QComboBox.

Signed-off-by: Dimitri Papadopoulos <3234522+DimitriPapadopoulos@users.noreply.github.com>