Address Linux crash on startup
Currently since many releases the application does not run on
Linux. That's because of our use of isSystemTrayAvailable(). On
Windows and MacOSX this function is a no-op (always true), while
on Linux there are a few steps that need to be executed. For that
we should have called it after app is setup and not in a constructor.
This patch moves the call prior to application setup.

Signed-off-by: Nikos Mavrogiannopoulos <n.mavrogiannopoulos@gmail.com>