MainWindow: Allow for immediate access to connecting to server
The main user task in the main window is to connect to a server.
Currently we force the user to go through create a new profile
form before they can proceed to their task.

This commit eliminates this intermediate step and allows
the user to immediately type the server to connect to. That
way we connect the main user task with the window's purpose
even for new connections.

Resolves: #342

Signed-off-by: Nikos Mavrogiannopoulos <n.mavrogiannopoulos@gmail.com>